requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
fsspec==0.8.7
future==0.18.2
grpcio==1.36.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
markdown==3.3.4
matplotlib==3.3.4
more-itertools==8.7.0
multidict==5.1.0
oauthlib==3.1.0
pandas==1.2.3
protobuf==3.15.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.3
pytz==2021.1
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2
scikit-plot==0.3.7
scipy==1.6.1
six==1.15.0
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1
threadpoolctl==2.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.3
werkzeug==1.0.1
yarl==1.6.3
numpy==1.20.1
torch==1.8.1
Pillow==8.1.2
scikit_learn==0.24.1
timm==0.4.5
torchvision==0.9.1
